+++ src/f_pc/searcher.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <SComponent.h>
+#include <unistd.h>
 
 namespace f_pc_searcher {
-	int fpcSch_JudgeByID(base_process_class *param_1, int *param_2);
+	intptr_t fpcSch_JudgeByID(base_process_class *param_1, int *param_2);
 }
